fully remove all code related to sending data offsite
This is a step past the first commit, and was my intention, though I
recognize people may actually want to use the 'feature' provided.

However, this is what I think it _should_ do. This is a useless feature
that hurts you in multiple ways over multiple timescales. Not only does
it expose you to potential breach (what if I had sensitive info I
committed accidentally, locally, and devenv grabs it and shoots it to
some server? What if I have a private repo where I _deliberately place_
such information? What if I work on PII? What if I work on
safety-critical software with vulnerabilities that now can be observed
by some guy's 'AI' -- whatever the heck that might be?

This is not the only reason this braindead, myopic feature should be
disabled. 'Generating' a devshell is about the _stupidest_ thing you
could do. Not only do you now not know what your shell will contain, but
you don't even know how to configure it for yourself.

I'm not against the use of AI unilaterally but this is a very dumb way
of integrating it.